% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-tensor.R
\name{as_tf_tensor}
\alias{as_tf_tensor}
\alias{as_tf_tensor.tf_tensor}
\alias{as_tf_tensor.array}
\alias{tf_tensor_clone}
\alias{tf_tensor_from_bytes}
\alias{tf_tensor_to_bytes}
\alias{as.array.tf_tensor}
\alias{ptype_from_tf_tensor}
\alias{tf_ptype}
\alias{tf_ptype_from_label}
\title{Create TensorFlow Tensors}
\usage{
as_tf_tensor(x, ...)

\method{as_tf_tensor}{tf_tensor}(x, ...)

\method{as_tf_tensor}{array}(x, ..., .tf_ptype = tf_ptype(x))

tf_tensor_clone(x)

tf_tensor_from_bytes(bytes, dim, .tf_ptype)

tf_tensor_to_bytes(x)

\method{as.array}{tf_tensor}(x, ..., .ptype = ptype_from_tf_tensor(x))

ptype_from_tf_tensor(x)

tf_ptype(x)

tf_ptype_from_label(label)
}
\arguments{
\item{x}{An object to convert to a tensor}

\item{...}{Unused}

\item{.tf_ptype}{A data type identifier, probably generated by \code{\link[=tf_ptype]{tf_ptype()}}
or \code{\link[=tf_ptype_from_label]{tf_ptype_from_label()}}.}

\item{bytes}{The uninterpreted underlying data as a \code{raw()} vector.}

\item{dim}{A vector of dimensions used to create the tensor}

\item{.ptype}{An R vector template (e.g., \code{\link[=double]{double()}}, \code{\link[=integer]{integer()}}, \code{\link[=logical]{logical()}},
\code{\link[=raw]{raw()}}, \code{\link[=complex]{complex()}}, or \code{\link[=character]{character()}}).}

\item{label}{One of 'FLOAT', 'DOUBLE', 'INT32', 'UINT8', 'INT16', 'INT8',
'STRING', 'COMPLEX64', 'INT64', 'BOOL', 'UINT16', 'COMPLEX128', 'UINT32',
or 'UINT64'.}
}
\value{
An object of class 'tf_tensor'
}
\description{
Create TensorFlow Tensors
}
\examples{
(t <- as_tf_tensor(matrix(1)))
tf_tensor_attributes(t)

}
