% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BassCurveAnalogous.R
\name{Bass_AC}
\alias{Bass_AC}
\title{Analogous Bass Curve}
\usage{
Bass_AC(p, q, m, t)
}
\arguments{
\item{p}{is a number represents coefficient of innovation.}

\item{q}{ia a number represents coefficient of immitation.}

\item{m}{is a number represents maximum market size can be reached.}

\item{t}{is an integer shows the growth periods.}
}
\value{
data frame (period, adoption rate, cumulative adoption)
}
\description{
Analogous Bass Curve uses analogy to estimate Bass Curve parameters. In case that there is no
data available for technology, parameters of similar cases can be used to forecast growth curve.
}
\examples{
Bass_AC(0.016, 0.304, 100000, 20)
\dontrun{

}
}

