% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CATV}
\alias{CATV}
\title{Growth of Cable TV Industriy}
\format{dataframe with two variables:
\describe{
\item{year}{year}
\item{CATV}{Cable TV cumulative adoption}}}
\usage{
CATV
}
\description{
This data shows the Cable TV adoption 1952-1989.
}
\references{
J. P. Martino, Technological forecasting for decision making. McGraw-Hill, Inc., 1993.
}
\keyword{datasets}

