% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalRepresentation.R
\name{cleaning}
\alias{cleaning}
\title{Deletion of location that are not in New york and its suburbs}
\usage{
cleaning(train)
}
\arguments{
\item{train}{object of class dataset}
}
\value{
A new dataframe with locations of interest as observation
}
\description{
Deletion of location that are not in New york and its suburbs
}
