% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.data.frame.tf}
\alias{as.data.frame.tf}
\alias{as.matrix.tfd}
\alias{as.matrix.tfb}
\alias{as.function.tf}
\title{Convert functional data back to tabular data formats}
\usage{
\method{as.data.frame}{tf}(x, row.names = NULL, optional = FALSE, unnest = FALSE, ...)

\method{as.matrix}{tfd}(x, arg, interpolate = FALSE, ...)

\method{as.matrix}{tfb}(x, arg = tf_arg(x), ...)

\method{as.function}{tf}(x, ...)
}
\arguments{
\item{x}{a \link{tfb} object to be converted}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{not used!}

\item{unnest}{if \code{TRUE}, the function will return a data.frame with the evaluated functions.}

\item{...}{additional arguments to be passed to or from methods.}

\item{arg}{a grid of argument values to evaluate on.}
}
\value{
\strong{for \code{as.data.frame.tf}:} if \code{unnest} is \code{FALSE} (default), a one-column \code{data.frame} with a \code{tf}-column containing \code{x}.
if \code{unnest} is \code{TRUE}, a 3-column data frame with columns \code{id}
for the (unique) names of \code{x} or a numeric identifier, \code{arg} and \code{value},
with each row containing one function evaluation at the original \code{arg}-values.
}
\description{
Various converters to turn \code{tfb}- or \code{tfd}-vectors back into data.frames or
matrices (or actual functions, even!).
}
