% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxiModel.R
\name{discretisation_dataframe}
\alias{discretisation_dataframe}
\title{Whole dataframe error or final data}
\usage{
discretisation_dataframe(good_df)
}
\arguments{
\item{good_df}{a dataframe which went through transformation}
}
\value{
A whole dataframe containing errors or the final data
}
\description{
Whole dataframe error or final data
}
