% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GompertzCurveAnalogous.R
\name{Gompertz_AC}
\alias{Gompertz_AC}
\title{Analogous Gompertz Curve}
\usage{
Gompertz_AC(b, k, l, t)
}
\arguments{
\item{b}{is a number represents curve coefficients.}

\item{k}{is a number represents curve coefficients.}

\item{l}{is a number represents maximum growth can be reached.}

\item{t}{is an integer shows the growth periods.}
}
\value{
data frame (period, cumulative adoption)
}
\description{
Analogous Gompertz Curve uses analogy to determine curve coefficients (b & k) .
}
\examples{
\dontrun{
Gompertz_AC(0.8, 0.2, 10000, 20)
}
}

