% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GompertzCurveAnalogous.R
\name{Gompertz_AC_Plot}
\alias{Gompertz_AC_Plot}
\title{Analogous Gompertz Curve Plot}
\usage{
Gompertz_AC_Plot(b, k, l, t)
}
\arguments{
\item{b}{is a number represents curve coefficients.}

\item{k}{is a number represents curve coefficients.}

\item{l}{is a number represents maximum growth can be reached.}

\item{t}{is an integer shows the growth periods.}
}
\value{
plot of cumulative adoption over certain period of time.
}
\description{
Analogous Gompertz Curve Plot function plots Analogous Gompertz Curve.
}
\examples{
Gompertz_AC_Plot(0.8, 0.2, 10000, 20)
\dontrun{

}
}

