% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GompertzCurveHistorical.R
\name{Gompertz_HC}
\alias{Gompertz_HC}
\title{Historical Gompertz Curve}
\usage{
Gompertz_HC(x, l, t)
}
\arguments{
\item{x}{is data frame with two columns to estimates b & k parameters. First column contains the period and second column contains
the associated cumulative adoption.}

\item{l}{is a number represents maximum growth can be reached.}

\item{t}{is an integer shows the growth periods.}
}
\value{
data frame (period, best fitted cumulative adoption)
}
\description{
Historical Gomperz Curve function estimates the curve parameters (b & k) based on the historical data.
}
\examples{
data("PE")
Gompertz_HC(PE, 2.81, 20)
\dontrun{

}
}

