% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GompertzCurveHistorical.R
\name{Gompertz_HC_plot}
\alias{Gompertz_HC_plot}
\title{Historical Gompertz Curve Plot}
\usage{
Gompertz_HC_plot(x, l, t)
}
\arguments{
\item{x}{is data frame with two columns to estimates b & k parameters. First column contains the period and second column contains
the associated cumulative adoption.}

\item{l}{is a number represents maximum growth can be reached.}

\item{t}{is an integer shows the growth periods.}
}
\value{
plot of best fitted cumulative adoption.
}
\description{
Historical Gomperz Curve Plot function plots Historical Gomperz Curve function results.
}
\examples{
\dontrun{
data("PE")
Gompertz_HC_plot(PE, 2.81, 250)
}
}

