% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxiModel.R
\name{good_dataframe}
\alias{good_dataframe}
\title{Whole dataframe with clean data}
\usage{
good_dataframe(df, grid)
}
\arguments{
\item{df}{The dataframe containing all the initial information}

\item{grid}{The spatial grid we chose to work with}
}
\value{
creates a whole dataframe with clean data (we apply transformation on all the lines)
}
\description{
Whole dataframe with clean data
}
