% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalRepresentation.R
\name{main}
\alias{main}
\title{Return the Spatial Grid that will serve as the domain}
\usage{
main()
}
\value{
The spatial grid from scratch to save it in our data folder to be used in the model side
}
\description{
Return the Spatial Grid that will serve as the domain
}
