% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalRepresentation.R
\name{over_2}
\alias{over_2}
\title{Return Id from our grid given latitude and longitude inputs}
\usage{
over_2(long, lat, grid)
}
\arguments{
\item{long}{double representing the longitude}

\item{lat}{double representing the latitude}

\item{grid}{the spatialgrid of interest}
}
\value{
Id from our grid
}
\description{
Return Id from our grid given latitude and longitude inputs
}
