% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxiModel.R
\name{path}
\alias{path}
\title{Final dataset for one trip}
\usage{
path(pretty_row)
}
\arguments{
\item{pretty_row}{A row given by the function "transformation"}
}
\value{
A data frame line which corresponds to a trip and contains :
         ID, passenger, day of the week, time slot, precision and price
}
\description{
Final dataset for one trip
}
