% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PE}
\alias{PE}
\title{Power-Generation Efficiency of Public Utilities}
\format{dataframe with two variables:
\describe{
\item{year}{year}
\item{Kwhperc}{Kilowatt.hours per pound of coal}}}
\usage{
PE
}
\description{
This data shows the efficiency improvement 1920 - 1987.
}
\references{
J. P. Martino, Technological forecasting for decision making. McGraw-Hill, Inc., 1993.
}
\keyword{datasets}

