% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PearlCurveAnalogous.R
\name{Pearl_AC}
\alias{Pearl_AC}
\title{Analogous pearl Curve}
\usage{
Pearl_AC(a, b, l, t)
}
\arguments{
\item{a}{is a number represents curve coefficients.}

\item{b}{is a number represents curve coefficients.}

\item{l}{is a number represents maximum growth can be reached.}

\item{t}{is an integer shows the growth periods.}
}
\value{
data frame (period, cumulative adoption)
}
\description{
Analogous Pearl Curve uses analogy to determine curve coefficients (a & b) .
}
\examples{
Pearl_AC(1.6, 0.8, 10000, 20)
\dontrun{

}
}

