% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PearlCurveAnalogous.R
\name{Pearl_AC_Plot}
\alias{Pearl_AC_Plot}
\title{Analogous pearl Curve Plot}
\usage{
Pearl_AC_Plot(a, b, l, t)
}
\arguments{
\item{a}{is a number represents curve coefficients.}

\item{b}{is a number represents curve coefficients.}

\item{l}{is a number represents maximum growth can be reached.}

\item{t}{is an integer shows the growth periods.}
}
\value{
plot of cumulative adoption over certain period of time.
}
\description{
Analogous Pearl Curve Plot function plots Analogous Pearl Curve.
}
\examples{
\dontrun{
Pearl_AC_Plot(1.6, 0.8, 10000, 20)
}
}

