% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PearlCurveHistorical.R
\name{Pearl_HC}
\alias{Pearl_HC}
\title{Historical Pearl Curve}
\usage{
Pearl_HC(x, l, t)
}
\arguments{
\item{x}{is data frame with two columns to estimates b & k parameters. First column contains the period and second column contains
the associated cumulative adoption.}

\item{l}{is a number represents maximum growth can be reached.}

\item{t}{is an integer shows the growth periods.}
}
\value{
data frame (period, best fitted cumulative adoption)
}
\description{
Historical Pearl Curve function estimates the curve parameters (a & b) based on the historical data.
}
\examples{
\dontrun{
data("CATV")
Pearl_HC(CATV,89024390,20)
}
}

