% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PearlCurveHistorical.R
\name{Pearl_HC_plot}
\alias{Pearl_HC_plot}
\title{Historical Pearl Curve Plot}
\usage{
Pearl_HC_plot(x, l, t)
}
\arguments{
\item{x}{is data frame with two columns to estimates b & k parameters. First column contains the period and second column contains
the associated cumulative adoption.}

\item{l}{is a number represents maximum growth can be reached.}

\item{t}{is an integer shows the growth periods.}
}
\value{
plot of best fitted cumulative adoption.
}
\description{
Historical Pearl Curve Plot function plots Historical Pearl Curve growth function results.
}
\examples{
\dontrun{
data("CATV")
Pearl_HC_plot(CATV,89024390,20)
}
}

