% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pinch}
\alias{pinch}
\title{Pinch force data}
\format{
An object of class \code{tfd_reg} (inherits from \code{tfd}, \code{tf}, \code{vctrs_vctr}, \code{list}) of length 20.
}
\usage{
pinch
}
\description{
Measurements of pinch force during 20 replications, with 151 observations
recorded every 2 milliseconds over 300 milliseconds. The data captures the
dynamics of finger pinch force applied during controlled motor tasks.
}
\details{
Data is also include in the \CRANpkg{fda} package in another format.
}
\examples{
pinch
}
\references{
Ramsay, O. J, Hooker, Giles, Graves, Spencer (2009).
\emph{Functional Data Analysis with R and MATLAB},  series Use R!, 1 edition.
Springer New York, New York.
ISBN 978-0-387-98184-0, \doi{10.1007/978-0-387-98185-7}.

Ramsay, O. J, Silverman, W. B (2005).
\emph{Functional Data Analysis},  series Springer Series in Statistics, 2nd edition.
Springer, New York.
ISBN 978-0-387-40080-8.

Ramsay, O. J, Silverman, W. B (2002).
\emph{Applied Functional Data Analysis}.
Springer.
}
\keyword{datasets}
