% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxiModel.R
\name{predict}
\alias{predict}
\title{Prediction of taxi prices}
\usage{
predict(travel)
}
\arguments{
\item{travel}{Vector obtained from transform_row which contains the informaton about a trip we want to make
(coordinates, day, hour, passengers)}
}
\value{
The estimated price of the trip
}
\description{
Prediction of taxi prices
}
