% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{prep_plotting_arg}
\alias{prep_plotting_arg}
\title{Preprocess evaluation grid for plotting}
\usage{
prep_plotting_arg(f, n_grid)
}
\arguments{
\item{f}{a \code{tf}-object.}

\item{n_grid}{length of evaluation grid.}
}
\value{
a semi-regular grid rounded down to appropriate resolution.
}
\description{
(internal function exported for re-use in upstream packages)
}
\examples{
f <- tfd(sin(seq(0, 2 * pi, length.out = 21)), arg = seq(0, 1, length.out = 21))
prep_plotting_arg(f, n_grid = 50)
}
\seealso{
Other tidyfun developer tools: 
\code{\link{ensure_list}()},
\code{\link{unique_id}()}
}
\concept{tidyfun developer tools}
