% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalRepresentation.R
\name{proj_cord}
\alias{proj_cord}
\title{Create a plot}
\usage{
proj_cord(shp, train)
}
\arguments{
\item{shp}{shapefile}

\item{train}{the clean version from the cleaning function}
}
\value{
Useful result to plot our data on the good grid
}
\description{
Create a plot
}
