% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalRepresentation.R
\name{proj_shp}
\alias{proj_shp}
\title{Create a shapefile}
\usage{
proj_shp(url)
}
\arguments{
\item{url}{Information online about NYC}
}
\value{
the shapefile
}
\description{
Create a shapefile
}
