% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-package.R
\docType{package}
\name{tf-package}
\alias{tf}
\alias{tf-package}
\title{tf: 'TensorFlow' C API Wrapper}
\description{
Wraps the 'TensorFlow' C API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/tf}
  \item Report bugs at \url{https://github.com/paleolimbot/tf/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
