% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-package.R
\docType{package}
\name{tf}
\alias{tf}
\title{tf}
\description{
\code{tf} is the basic package prviding the infrastructure for \code{tidyfun} -- a
colleciton of packages that make data wrangling and exploratory analysis for
functional data in \code{R} easier.\cr\cr
\code{tf} provides:
}
\details{
\itemize{
\item new data types for representing functional data: \code{\link[=tfd]{tfd()}} & \code{\link[=tfb]{tfb()}}
\item arithmetic operators (\code{\link[=Ops.tf]{Ops.tf()}}),
\item descriptive statistics: e.g. \code{\link[=mean.tf]{mean.tf()}}, \code{\link[=median.tf]{median.tf()}}
\item base graphics functions for such data: \code{\link[=plot.tf]{plot.tf()}}
\item functions to do smoothing (\code{\link[=tf_smooth.tfd]{tf_smooth.tfd()}}), differentiation and integration (\code{\link[=tf_derive.tfd]{tf_derive.tfd()}})
\cr\cr
\cr
}
}
