% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{tf_align}
\alias{tf_align}
\alias{tf_align.tfd}
\alias{tf_align.tfb}
\title{Apply warping functions to align functional data}
\usage{
tf_align(x, warp, ...)

\method{tf_align}{tfd}(x, warp, ..., keep_new_arg = FALSE)

\method{tf_align}{tfb}(x, warp, ...)
}
\arguments{
\item{x}{\code{tf} vector of functions. For \code{tf_warp()}, these should be
registered/aligned functions and unaligned functions for \code{tf_align()}.}

\item{warp}{\code{tf} vector of warping functions used for transformation. See Details.}

\item{...}{additional arguments passed to \code{\link[=tfd]{tfd()}}.}

\item{keep_new_arg}{keep new \code{arg} values after (un)warping or return
\code{tfd} vector on \code{arg} values of the input (default \code{FALSE} is the latter)? See Details.}
}
\value{
the aligned \code{tf} vector (registered functions)
}
\description{
\code{tf_align()} applies the \emph{inverse} warping function to unregistered data
to obtain aligned (registered) functions.
}
\examples{
# Estimate warps, then align manually:
t <- seq(0, 2 * pi, length.out = 101)
x <- tfd(t(sapply(c(-0.3, 0, 0.3), function(s) sin(t + s))), arg = t)
warps <- tf_estimate_warps(x, method = "affine", type = "shift")
aligned <- tf_align(x, warps)
plot(aligned, col = 1:3)
}
\seealso{
Other registration functions: 
\code{\link{tf_estimate_warps}()},
\code{\link{tf_landmarks_extrema}()},
\code{\link{tf_register}()},
\code{\link{tf_registration}},
\code{\link{tf_warp}()}
}
\concept{registration functions}
