% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv-methods.R
\name{tf_arclength}
\alias{tf_arclength}
\alias{tf_arclength.default}
\alias{tf_arclength.tf_mv}
\title{Arc length of vector-valued functional data}
\usage{
tf_arclength(f, ...)

\method{tf_arclength}{default}(f, ...)

\method{tf_arclength}{tf_mv}(f, arg, lower, upper, definite = TRUE, ...)
}
\arguments{
\item{f}{a \code{tf_mv} object.}

\item{...}{forwarded to \code{\link[=tf_integrate]{tf_integrate()}}.}

\item{arg, lower, upper}{see \code{\link[=tf_integrate]{tf_integrate()}}.}

\item{definite}{\code{TRUE} (default) returns one total arc length per curve (a
numeric vector); \code{FALSE} returns the cumulative arc length \eqn{s(t) =
  \int_a^t \lVert f'(u) \rVert\, du} as a univariate \code{tfd}.}
}
\value{
a numeric vector (definite) or a univariate \code{tfd} (indefinite).
}
\description{
For a vector-valued curve \verb{f: [a, b] -> R^d}, the arc length is
\eqn{\int_a^b \lVert f'(t) \rVert\, dt}, i.e. the length traced out by
\code{f} in \code{R^d}. This implementation simply composes the existing univariate
verbs: per-component differentiation (\code{\link[=tf_derive]{tf_derive()}}), pointwise Euclidean
norm of the derivative (\verb{sqrt(sum_k f'_k^2)}), then numerical integration
(\code{\link[=tf_integrate]{tf_integrate()}}).
}
\examples{
# unit circle parameterised on [0, 1] -- arc length should be ~ 2*pi
t <- seq(0, 1, length.out = 201)
x <- tfd(matrix(cos(2 * pi * t), nrow = 1), arg = t)
y <- tfd(matrix(sin(2 * pi * t), nrow = 1), arg = t)
tf_arclength(tfd_mv(list(x = x, y = y)))
}
\seealso{
Other tf_mv-class: 
\code{\link{plot.tf_mv}()},
\code{\link{tf_ncomp}()},
\code{\link{tfb_mv}},
\code{\link{tfd_mv}}
}
\concept{tf_mv-class}
