% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{tf_depth}
\alias{tf_depth}
\alias{tf_depth.matrix}
\alias{tf_depth.tf}
\title{Functional Data Depth}
\usage{
tf_depth(x, arg, depth = "MBD", na.rm = TRUE, ...)

\method{tf_depth}{matrix}(x, arg, depth = c("MBD", "MEI"), na.rm = TRUE, ...)

\method{tf_depth}{tf}(x, arg, depth = "MBD", na.rm = TRUE, ...)
}
\arguments{
\item{x}{\code{tf} (or a matrix of evaluations).}

\item{arg}{grid of evaluation points.}

\item{depth}{currently available: \code{"MBD"}, i.e. modified 2-band depth, and \code{"MEI"}.}

\item{na.rm}{remove missing observations? Defaults to \code{TRUE}.}

\item{...}{further arguments handed to the function computing the respective
tf_depth.}
}
\value{
vector of tf_depth values
}
\description{
Data depths for functional data.
Currently implemented:
\itemize{
\item Modified Band-2 Depth
\item Modified Epigraph Index.
}
}
\details{
Roughly, modified band depth computes the centrality
of a function (scale of 0 (extreme) to 0.5 (central)), while the epigraph index computes how often it is above
other functions (scale of 0 (lowest) to 1 (highest)).
The two are closely related -- for functions that never cross other functions,
MBD is \eqn{-2(MEI - 0.5)^2 + .5}.
}
\references{
Sun, Ying, Genton, G M, Nychka, W D (2012).
\dQuote{Exact fast computation of band depth for large functional datasets: How quickly can one million curves be ranked?}
\emph{Stat}, \bold{1}(1), 68--74.

López-Pintado, Sara, Romo, Juan (2009).
\dQuote{On the concept of depth for functional data.}
\emph{Journal of the American Statistical Association}, \bold{104}(486), 718--734.

López-Pintado, Sara, Romo, Juan (2011).
\dQuote{A half-region depth for functional data.}
\emph{Computational Statistics & Data Analysis}, \bold{55}(4), 1679--1695.
}
\concept{tidyfun ordering and ranking functions}
