% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{tf_derive}
\alias{tf_derive}
\alias{tf_derive.matrix}
\alias{tf_derive.tfd}
\alias{tf_derive.tfd_irreg}
\alias{tf_derive.tfb_spline}
\alias{tf_derive.tfb_fpc}
\title{Differentiating functional data: approximating derivative functions}
\usage{
tf_derive(f, arg, order = 1, ...)

\method{tf_derive}{matrix}(f, arg, order = 1, ...)

\method{tf_derive}{tfd}(f, arg = tf_arg(f), order = 1, ...)

\method{tf_derive}{tfd_irreg}(f, arg, order = 1, ...)

\method{tf_derive}{tfb_spline}(f, arg = tf_arg(f), order = 1, ...)

\method{tf_derive}{tfb_fpc}(f, arg = tf_arg(f), order = 1, ...)
}
\arguments{
\item{f}{a \code{tf}-object}

\item{arg}{grid to use for the finite differences.
Not (exactly) the \code{arg} of the returned object for \code{tfd}-inputs, see details.}

\item{order}{order of differentiation. Maximal value for \code{tfb_spline} is 2.
For \code{tfb_spline}-objects, \code{order = -1} yields integrals (used internally).}

\item{...}{not used}
}
\value{
a \code{tf} (with slightly different \code{arg} or \code{basis} for the
derivatives, see details).
}
\description{
Derivatives of \code{tf}-objects use finite differences of the evaluations for
\code{tfd} and finite differences of the basis functions for \code{tfb}.
}
\details{
The derivatives of \code{tfd} objects use second-order accurate central differences
for interior points and second-order accurate one-sided differences at
boundaries, following the non-uniform grid formulas from \code{numpy.gradient}
with \code{edge_order=2} (Fornberg, 1988).
Domain and grid of the returned object are identical to the input. Unless the
\code{tfd} has a rather fine and regular grid, representing the data in a suitable
basis representation with \code{\link[=tfb]{tfb()}} and then computing the derivatives (or
integrals) of those is usually preferable.

Note that, for spline bases like \code{"cr"} or \code{"tp"} which are constrained to
begin/end linearly, computing \emph{second} derivatives will produce artefacts at
the outer limits of the functions' domain due to these boundary constraints.
Basis \code{"bs"} does not have this problem for sufficiently high orders (but
tends to yield slightly less stable fits).
}
\section{Methods (by class)}{
\itemize{
\item \code{tf_derive(matrix)}: row-wise finite differences

\item \code{tf_derive(tfd)}: derivatives by finite differencing of function evaluations.

\item \code{tf_derive(tfd_irreg)}: element-wise finite differencing for irregular grids.
Falls back to \code{tf_derive.tfd} (interpolating to a common grid) if an
explicit \code{arg} vector is supplied.

\item \code{tf_derive(tfb_spline)}: derivatives by finite differencing of spline basis functions.

\item \code{tf_derive(tfb_fpc)}: derivatives by finite differencing of FPC basis functions.

}}
\examples{
arg <- seq(0, 1, length.out = 31)
x <- tfd(rbind(arg^2, sin(2 * pi * arg)), arg = arg)
dx <- tf_derive(x)
x
dx
tf_arg(dx)
}
\seealso{
Other tidyfun calculus functions: 
\code{\link{tf_integrate}()}
}
\concept{tidyfun calculus functions}
