% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tf_fashion_mnist_test_images}
\alias{tf_fashion_mnist_test_images}
\title{Test images for the Fashion-MNIST}
\format{
An object of class \code{array} of dimension 100 x 28 x 28.
}
\usage{
tf_fashion_mnist_test_images
}
\description{
100 test images from the Fashion-MNIST data set as provided by
\code{keras::dataset_fashion_mnist()}.
}
\keyword{datasets}
