% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{tf_invert}
\alias{tf_invert}
\title{Invert a tf vector}
\usage{
tf_invert(x)
}
\arguments{
\item{x}{a tf vector.}
}
\value{
a tf vector of the inverted functions.
}
\description{
Computes the functional inverse of each function in the tf vector, such that
if \eqn{y = f(x)}, then \eqn{x = f^{-1}(y)}.
}
\examples{
arg <- seq(0, 2, length.out = 50)
x <- tfd(rbind(2 * arg, arg^2), arg = arg)
x_inv <- tf_invert(x)
layout(t(1:2))
plot(x, main = "original functions", ylab = "")
plot(x_inv, main = "inverted functions", ylab = "")
}
