% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{tf_jiggle}
\alias{tf_jiggle}
\alias{tf_jiggle.tf_mv}
\alias{tf_sparsify}
\alias{tf_sparsify.tf_mv}
\title{Make a \code{tf} (more) irregular}
\usage{
tf_jiggle(f, amount = 0.4, ...)

\method{tf_jiggle}{tf_mv}(f, amount = 0.4, same_arg = TRUE, ...)

tf_sparsify(f, dropout = 0.5, ...)

\method{tf_sparsify}{tf_mv}(f, dropout = 0.5, same_arg = TRUE, ...)
}
\arguments{
\item{f}{a \code{tfd} object.}

\item{amount}{how far away from original grid points can the jiggled grid points
lie, at most (relative to original distance to neighboring grid points).
Defaults to at most 40\% (0.4) of the original grid distances. Must be lower
than 0.5.}

\item{...}{additional args for the returned \code{tfd} in \code{tf_jiggle}.}

\item{same_arg}{for \code{tf_mv} objects, should all components receive the same
random argument-grid changes? Defaults to \code{TRUE}; use \code{FALSE} to jitter or
sparsify each component independently.}

\item{dropout}{what proportion of values of \code{f} to drop, on average. Defaults to half.}
}
\value{
an (irregular) \code{tfd} object.
}
\description{
Randomly create some irregular functional data from regular ones.
\itemize{
\item \strong{jiggle} it by randomly moving around its \code{arg}-values inside the intervals defined by its grid neighbors on the original argument grid.
\item \strong{sparsify} it by removing (100*\code{dropout})\% of the function values
}
}
\examples{
set.seed(1)
(x <- tf_rgp(2, arg = 21L))
(x_jig <- tf_jiggle(x, amount = 0.2))
(x_sp <- tf_sparsify(x, dropout = 0.3))
c(is_irreg(x_jig), is_irreg(x_sp))
}
\seealso{
Other tidyfun RNG functions:
\code{\link[=tf_rgp]{tf_rgp()}}
}
\concept{tidyfun RNG functions}
