% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-session.R
\name{tf_load_session_from_saved_model}
\alias{tf_load_session_from_saved_model}
\title{Title}
\usage{
tf_load_session_from_saved_model(export_dir, tags)
}
\arguments{
\item{export_dir}{A directory to a saved model}

\item{tags}{must include the set of tags used to identify one MetaGraphDef in
the SavedModel.}
}
\value{
An object of class 'tf_session'
}
\description{
Title
}
\examples{
saved <- system.file("extdata/fashion_mnist", package = "tf")
session <- tf_load_session_from_saved_model(saved, "serve")

}
