% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv-methods.R
\name{tf_ncomp}
\alias{tf_ncomp}
\alias{tf_components}
\alias{tf_component}
\alias{tf_component<-}
\title{Accessors and methods for vector-valued functional data}
\usage{
tf_ncomp(f)

tf_components(f)

tf_component(f, which)

tf_component(f, which) <- value
}
\arguments{
\item{f, x}{a \code{tf_mv} object.}

\item{which}{a component name or index.}

\item{value}{a univariate \code{tf} vector (replacement) of matching length and
domain.}
}
\value{
\code{tf_ncomp()}: an integer. \code{tf_components()}: a named list of \code{tf}
vectors. \code{tf_component()}: a single univariate \code{tf} vector.
}
\description{
Utilities for \code{tf_mv} objects (see \code{\link[=tfd_mv]{tfd_mv()}} / \code{\link[=tfb_mv]{tfb_mv()}}). \code{tf_ncomp()}
returns the number of output dimensions \eqn{d}, \code{tf_components()} the list
of the \code{d} underlying univariate \code{tf} vectors, and \code{tf_component()} extracts
or replaces a single one (also available via the \code{$} operator, e.g. \code{f$x}).
}
\examples{
f <- tfd_mv(list(x = tf_rgp(3), y = tf_rgp(3)))
tf_ncomp(f)
tf_components(f)
tf_component(f, "y")
f$y
}
\seealso{
Other tf_mv-class: 
\code{\link{plot.tf_mv}()},
\code{\link{tfb_mv}},
\code{\link{tfd_mv}}
}
\concept{tf_mv-class}
