% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{tf_register}
\alias{tf_register}
\title{Register a tf vector}
\usage{
tf_register(x, arg, verbose = FALSE, ...)
}
\arguments{
\item{x}{a tf vector to register.}

\item{arg}{the arg vector.}

\item{verbose}{print progress messages?}
}
\value{
\code{list()} containing: registered tf vector \code{x} and warping functions \code{warp}.
}
\description{
Register a tf vector
}
