% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{tf_rgam}
\alias{tf_rgam}
\title{Generate random warping functions}
\usage{
tf_rgam(n, arg = 51L, sigma = 0.1)
}
\arguments{
\item{n}{number of random warping functions to generate.}

\item{arg}{number of grid points per curve.}

\item{sigma}{standard deviation of the Gaussian noise added to the warping
functions.}
}
\value{
tf vector of random warping functions.
}
\description{
Generate random warping functions
}
