% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-session.R
\name{tf_session_new}
\alias{tf_session_new}
\alias{tf_session_close}
\alias{tf_load_session_from_saved_model}
\alias{tf_session_graph}
\alias{tf_session_run}
\title{TensorFlow Sessions}
\usage{
tf_session_new(graph)

tf_session_close(x)

tf_load_session_from_saved_model(export_dir, tags)

tf_session_graph(x)

tf_session_run(x, input_operation, output_operation, input)
}
\arguments{
\item{graph}{A \link[=tf_graph_new]{tf_graph}}

\item{x}{An object of class 'tf_session'}

\item{export_dir}{A directory to a saved model}

\item{tags}{must include the set of tags used to identify one MetaGraphDef in
the SavedModel.}

\item{input_operation, output_operation}{Names of the input/output operations
from the \code{\link[=tf_session_graph]{tf_session_graph()}}.}

\item{input}{A \code{list()} of \link[=as_tf_tensor]{tf_tensor} or objects that can be
converted to then (notably, arrays).}
}
\value{
An object of class 'tf_session'
}
\description{
TensorFlow Sessions
}
\examples{
saved <- system.file("extdata/fashion_mnist.zip", package = "tf")
session <- tf_load_session_from_saved_model(saved, "serve")
(graph <- tf_session_graph(session))
tf_graph_list_operations(graph)

# run a prediction!
input <- as_tf_tensor(
  tf_fashion_mnist_test_images[1, , , drop = FALSE],
  .tf_ptype = "FLOAT"
)

result <- tf_session_run(
  session,
  "serving_default_flatten_input",
  "StatefulPartitionedCall",
  list(input)
)

result
as.array(result[[1]])
tf_session_close(session)

}
