% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-tensor.R
\name{tf_tensor_attributes}
\alias{tf_tensor_attributes}
\alias{tf_tensor_valid}
\title{Inspect Tensors}
\usage{
tf_tensor_attributes(x)

tf_tensor_valid(x)
}
\arguments{
\item{x}{A \link[=as_tf_tensor]{tf_tensor}}
}
\description{
Inspect Tensors
}
