% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{tf_unwarp}
\alias{tf_unwarp}
\title{Unwarp a tf vector}
\usage{
tf_unwarp(x, warp, ..., keep_arg = FALSE)
}
\arguments{
\item{x}{tf vector of unregistered functions.}

\item{warp}{tf vector of warping functions.}

\item{...}{additional arguments passed to \code{\link[=tfd]{tfd()}}.}

\item{keep_arg}{re-eval on original arg after warping or return (irregular)
tf on warped arg (default)?}
}
\value{
the unwarped tf vector, i.e. the registered functions.
}
\description{
Apply inverse of given warping functions to functional data:
\eqn{x(t) \to x(h^{-1}(t)) = x(s)}.
}
\examples{
arg <- seq(0, 2 * pi, length.out = 100)
reg <- rep(tfd(sin(arg), arg = arg), 3)
# create warping functions with varying phase shifts
warp <- tfd(unname(rbind(arg, arg + 0.3 * arg, arg + 0.1 * arg^1.5)), arg = arg)
unreg <- tf_warp(reg, warp)
# apply unwarping to register the functions back to common time scale
reg <- tf_unwarp(unreg, warp)
plot(reg)
}
