% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{tf_warp}
\alias{tf_warp}
\title{Warp a tf vector}
\usage{
tf_warp(x, warp, ..., keep_arg = FALSE)
}
\arguments{
\item{x}{tf vector of (registered) functions.}

\item{warp}{tf vector of warping functions.}

\item{...}{additional arguments passed to \code{\link[=tfd]{tfd()}}.}

\item{keep_arg}{re-evaluate on warped arg values after un-warping or return
tf on un-warped arg-vals (default)?}
}
\value{
the warped tf vector, i.e. the unregistered functions.
}
\description{
Apply given warping functions to aligned functional data to get back to the
original unaligned data: \eqn{x(s) \to x(h(s)) = x(t)}.
}
\examples{
arg <- seq(0, 2 * pi, length.out = 100)
x <- rep(tfd(sin(arg), arg = arg), 3)
# create warping functions with varying phase shifts
warp <- tfd(unname(rbind(arg, arg + 0.3 * arg, arg + 0.1 * arg^1.5)), arg = arg)
# apply warping to get phase-shifted sine waves
unreg <- tf_warp(x, warp)
plot(unreg)
}
