% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brackets.R
\name{tfbrackets}
\alias{tfbrackets}
\alias{[.tf}
\alias{index.tf}
\alias{[<-.tf}
\title{Accessing, evaluating, subsetting and subassigning \code{tf} vectors}
\usage{
\method{[}{tf}(x, i, j, interpolate = TRUE, matrix = TRUE)

\method{[}{tf}(x, i) <- value
}
\arguments{
\item{x}{an \code{tf}}

\item{i}{index of the observations (\code{integer}ish, \code{character} or \code{logical},
usual R rules apply)}

\item{j}{The \code{arg} used to evaluate the functions. A (list of) \code{numeric}
vectors.}

\item{interpolate}{should functions be evaluated (i.e., inter-/extrapolated)
for values in \code{arg} for which no original data is available? Only relevant for
the raw data class \code{tfd}, for which it defaults to \code{TRUE}. Basis-represented
\code{tfb} are always "interpolated".}

\item{matrix}{should the result be returned as a \code{matrix} or as a list of
\code{data.frame}s? If \code{TRUE}, \code{j} has to be a (list of a) single vector of
\code{arg}. See return value.}

\item{value}{\code{tf} object for subassignment. This is (currently) very strictly typed,
i.e. only objects that are of the same class and have compatible \code{arg} can be
subassigned.}
}
\value{
If \code{j} is missing, a subset of the functions in \code{x} as given by
\code{i}.\cr If \code{j} is given and \code{matrix == TRUE}, a numeric matrix of function
evaluations in which each row represents one function and each column
represents one \code{argval} as given in argument \code{j}, with an attribute
\code{arg}=\code{j} and row- and column-names derived from \code{x[i]} and \code{j}.\cr If
\code{j} is given and \code{matrix == FALSE}, a list of \code{tbl_df}s with columns
\code{arg} = \code{j} and \code{value} = evaluations at \code{j} for each observation in
\code{i}.
}
\description{
These functions access, subset, replace and evaluate \code{tf} objects.
For more information on creating \code{tf} objects and converting them to/from
\code{list}, \code{data.frame} or \code{matrix}, see \code{\link[=tfd]{tfd()}} and \code{\link[=tfb]{tfb()}}. See Details.\cr
}
\details{
Note that these break certain (terrible) R conventions for vector-like objects:\cr
\itemize{
\item no argument recycling,
\item no indexing with \code{NA},
\item no indexing with names not present in \code{x},
\item no indexing with integers \verb{> length(x)}
}

All of the above will trigger errors. Sub-assigning new elements to positions
beyond the original vector length still works and will fill up the missing elements
in between with \code{NAs}. This package was developed by fickle, irridescently
rainbow-colored unicorns.
}
