% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-format.R
\name{print.tf}
\alias{print.tf}
\alias{print.tfd_reg}
\alias{print.tfd_irreg}
\alias{print.tfb}
\alias{format.tf}
\title{Pretty printing and formatting for functional data}
\usage{
\method{print}{tf}(x, n = 5, ...)

\method{print}{tfd_reg}(x, n = 5, ...)

\method{print}{tfd_irreg}(x, n = 5, ...)

\method{print}{tfb}(x, n = 5, ...)

\method{format}{tf}(
  x,
  digits = 2,
  nsmall = 0,
  width = options()$width,
  sparkline = TRUE,
  prefix = FALSE,
  ...
)
}
\arguments{
\item{x}{any \R object (conceptually); typically numeric.}

\item{n}{how many elements of \code{x} to print out at most}

\item{...}{handed over to \code{\link[=format.tf]{format.tf()}}}

\item{digits}{a positive integer indicating how many significant digits
    are to be used for
    numeric and complex \code{x}.  The default, \code{NULL}, uses
    \code{\link[base]{getOption}("digits")}.  This is a suggestion: enough decimal
    places will be used so that the smallest (in magnitude) number has
    this many significant digits, and also to satisfy \code{nsmall}.
    (For more, notably the interpretation for complex numbers see
    \code{\link[base]{signif}}.)}

\item{nsmall}{the minimum number of digits to the right of the decimal
    point in formatting real/complex numbers in non-scientific formats.
    Allowed values are \code{0 <= nsmall <= 20}.}

\item{width}{\code{default} method: the \emph{minimum} field width or
    \code{NULL} or \code{0} for no restriction.

    \code{AsIs} method: the \emph{maximum} field width for non-character
    objects.  \code{NULL} corresponds to the default \code{12}.
  }

\item{sparkline}{use a sparkline representation? defaults to \code{TRUE}
(not available for irregular data)}

\item{prefix}{prefix with names / index positions? defaults to \code{FALSE}}
}
\value{
prints out \code{x} and returns it invisibly

a character representation of \code{x}
}
\description{
Prints and formats \code{tf}-objects for display. See Details / examples for options that
give finer control.
}
\details{
By default, \code{tf} objects on regular grids are shown as
\href{cli::spark_bar}{sparklines}, set \code{sparkline = FALSE} for a text
representation.

Sparklines are based on running mean values of the function values, but
these don't check for non-equidistant grids, so the visual impression will be
misleading for very unequal grid distances.

Sparklines use
\code{options()$width/3} bins for printing/formatting by default, use \code{bins}
argument to set the number of bins explicitly.
For \code{\link[tibble:reexports]{tibble::glimpse()}}, we use 8 bins by default for compact display.
}
\examples{
t <- seq(0, 1, l = 101)
cosine <- purrr::map(1:5, \(i) cos(2 * pi * i * t)) |> tfd(arg = t)
cosine
tf_sparsify(cosine, dropout = .8)

format(cosine, sparkline = FALSE)
format(cosine, bins = 5)
format(cosine, bins = 40)
#! very non-equidistant grids --> sparklines can mislead
cosine_2 <- tfd(cosine, arg = t^3)
format(cosine_2, bins = 40)
}
\concept{tidyfun print}
