% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{train_transfo2}
\alias{train_transfo2}
\title{dataset}
\format{dataframe with the result of our calculated model to be used in predict}
\usage{
train_transfo2
}
\description{
dataset
}
\keyword{datasets}
