% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxiModel.R
\name{transform_row}
\alias{transform_row}
\title{Data row to vector}
\usage{
transform_row(data_row)
}
\arguments{
\item{data_row}{a row from a data frame}
}
\value{
a vector containing the pickup and dropoff coordinates, week day, hour and passengers
}
\description{
Data row to vector
}
