% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxiModel.R
\name{transformation}
\alias{transformation}
\title{Keep the relevant variables}
\usage{
transformation(list, grid)
}
\arguments{
\item{list}{List coming from a line of our initial data frame, corresponds to a taxi ride.
list[1] is the date of the trip (in basic date format)
list[2] is the price of the trip
list[4],list[5] are the latitudes and longitudes of the pickup point
list[6],list[7] are the latitudes and longitudes of the dropoff point
list[8] is the ID of the passenger}

\item{grid}{Our spatial grid}
}
\value{
A new data frame with the variables which interest us :
        longitude of pickup
        latitude of pickup
        longitude of dropoff
        latitude of dropoff
        id of the pickup case
        id of the dropoff case
        the day of the week
        hour
        price
        passenger id
}
\description{
Keep the relevant variables
}
