% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-tensor.R
\name{as_tf_tensor}
\alias{as_tf_tensor}
\alias{as_tf_tensor.tf_tensor}
\alias{as_tf_tensor.array}
\alias{tf_tensor_clone}
\alias{as.array.tf_tensor}
\alias{tf_ptype_from_tensor}
\title{Create TensorFlow Tensors}
\usage{
as_tf_tensor(x, ...)

\method{as_tf_tensor}{tf_tensor}(x, ...)

\method{as_tf_tensor}{array}(x, ...)

tf_tensor_clone(x)

\method{as.array}{tf_tensor}(x, ..., .ptype = tf_ptype_from_tensor(x))

tf_ptype_from_tensor(x)
}
\arguments{
\item{x}{An object to convert to a tensor}

\item{...}{Unused}

\item{.ptype}{An R vector template (e.g., \code{\link[=double]{double()}}, \code{\link[=integer]{integer()}},
\code{\link[=raw]{raw()}}, \code{\link[=complex]{complex()}}, or \code{\link[=character]{character()}}).}
}
\value{
An object of class 'tf_tensor'
}
\description{
Create TensorFlow Tensors
}
\examples{
(t <- as_tf_tensor(matrix(1)))
tf_tensor_attributes(t)

}
