% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-tensor.R
\name{as_tf_tensor}
\alias{as_tf_tensor}
\alias{as_tf_tensor.array}
\title{Create TensorFlow Tensors}
\usage{
as_tf_tensor(x)

\method{as_tf_tensor}{array}(x)
}
\arguments{
\item{x}{An object to convert to a tensor}
}
\value{
An object of class 'tf_tensor'
}
\description{
Create TensorFlow Tensors
}
\examples{
(t <- as_tf_tensor(matrix(1)))
tf_tensor_attributes(t)

}
