% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_list}
\alias{ensure_list}
\title{Turns any object into a list}
\usage{
ensure_list(x)
}
\arguments{
\item{x}{any input}
}
\value{
\code{x} turned into a list.
}
\description{
See above.
}
\seealso{
Other tidyfun developer tools: 
\code{\link{prep_plotting_arg}()},
\code{\link{unique_id}()}
}
\concept{tidyfun developer tools}
