% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gait}
\alias{gait}
\title{Hip and knee angle while walking data}
\format{
A data frame with 39 rows and 3 variables:
\describe{
\item{subject_id}{subject identifier}
\item{knee_angle}{a \code{tfd} column of knee-joint angles (degrees)}
\item{hip_angle}{a \code{tfd} column of hip-joint angles (degrees)}
}
}
\usage{
gait
}
\description{
Hip and knee angle measurements in degrees through a 20-point movement cycle
for 39 children. The data represents the angular positions of hip and knee
joints during normal walking gait, captured at evenly spaced time points
throughout the gait cycle.
}
\details{
Data is also included in the \CRANpkg{fda} package in another format.
}
\examples{
head(gait)
}
\references{
Olshen, A R, Biden, N E, Wyatt, P M, Sutherland, H D (1989).
\dQuote{Gait Analysis and the Bootstrap.}
\emph{The Annals of Statistics}, \bold{17}(4), 1419--1440.
}
\keyword{datasets}
