% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gait}
\alias{gait}
\title{Hip and knee angle while walking data}
\format{
An object of class \code{data.frame} with 39 rows and 3 columns.
}
\usage{
gait
}
\description{
Hip and knee angle while walking data
}
\examples{
gait
}
\keyword{datasets}
