% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicalRepresentation.R
\name{grid}
\alias{grid}
\title{Title}
\usage{
grid(shp, largeur_cellule = 500)
}
\arguments{
\item{largeur_cellule}{Parameter to know how precise we cut our travel in our cells}

\item{shapefile}{Geographical object for NYC}
}
\value{
Final spatial grid used in our functions to create the data set model
}
\description{
Title
}
