% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth}
\alias{growth}
\title{Berkeley growth study data}
\format{
A data frame with 93 rows and 2 variables:
\describe{
\item{gender}{sex of the subject (boy/girl)}
\item{height}{height in centimeters}
}
}
\usage{
growth
}
\description{
Heights of 39 boys and 54 girls measured from age 1 to 18 years as part
of the Berkeley Growth Study. The data tracks physical development over
time with measurements at 31 different ages that are not equally spaced.'
}
\details{
Data is also include in the \CRANpkg{fda} package in another format.
}
\examples{
growth
}
\references{
Ramsay, O. J, Hooker, Giles, Graves, Spencer (2009).
\emph{Functional Data Analysis with R and MATLAB},  series Use R!, 1 edition.
Springer New York, New York.
ISBN 978-0-387-98184-0, \doi{10.1007/978-0-387-98185-7}.

Ramsay, O. J, Silverman, W. B (2005).
\emph{Functional Data Analysis},  series Springer Series in Statistics, 2nd edition.
Springer, New York.
ISBN 978-0-387-40080-8.

Ramsay, O. J, Silverman, W. B (2002).
\emph{Applied Functional Data Analysis}.
Springer.

Tuddenham, D R (1954).
\dQuote{Physical growth of California boys and girls from birth to eighteen years.}
\emph{University of California Publications in Child Development}, \bold{1}, 183--364.
}
\keyword{datasets}
