% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv-geom.R
\name{tf_geom}
\alias{tf_geom}
\alias{tf_norm}
\alias{tf_speed}
\alias{tf_inner}
\alias{tf_distance}
\alias{tf_tangent}
\alias{tf_reparam_arclength}
\title{Pointwise norm and inner product for vector-valued functional data}
\usage{
tf_norm(f)

tf_speed(f)

tf_inner(f, g)

tf_distance(f, g)

tf_tangent(f)

tf_reparam_arclength(f)
}
\arguments{
\item{f, g}{\code{tf_mv} objects (with identical \code{d} and component names where
two arguments are required).}
}
\value{
a univariate \code{tfd} for \code{tf_norm}/\code{tf_speed}/\code{tf_inner}/\code{tf_distance},
a \code{tf_mv} for \code{tf_tangent}/\code{tf_reparam_arclength}.
}
\description{
Small geometric helpers for \code{tf_mv} objects, defined by component-wise
composition of the existing univariate \code{Ops} / \code{Math} machinery:
\itemize{
\item \code{tf_norm(f)}     -- pointwise Euclidean norm \eqn{\lVert f(t) \rVert};
\item \code{tf_speed(f)}    -- pointwise speed \eqn{\lVert f'(t) \rVert};
\item \code{tf_inner(f, g)} -- pointwise inner product \eqn{\langle f(t), g(t) \rangle};
\item \code{tf_distance(f, g)} -- pointwise Euclidean distance \eqn{\lVert f(t) - g(t) \rVert};
\item \code{tf_tangent(f)}  -- unit tangent \eqn{f'(t) / \lVert f'(t) \rVert}, returned
as a \code{tf_mv} (undefined where the speed is zero -- callers get \code{NaN}s there);
\item \code{tf_reparam_arclength(f)} -- re-parametrize the curve at constant speed
(i.e. by its normalized cumulative arc length).
}
}
\examples{
set.seed(1)
f <- tfd_mv(list(x = tf_rgp(2), y = tf_rgp(2)))
tf_norm(f)
tf_speed(f)
tf_distance(f, tfd_mv(list(x = tf_rgp(2), y = tf_rgp(2))))
}
\seealso{
Other tf_mv-class:
\code{\link[=plot.tf_mv]{plot.tf_mv()}},
\code{\link[=tf_arclength]{tf_arclength()}},
\code{\link{tf_mv_methods}},
\code{\link[=tfb_mv]{tfb_mv()}},
\code{\link[=tfd_mv]{tfd_mv()}}
}
\concept{tf_mv-class}
