% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-graph.R
\name{tf_graph_import_graph_def}
\alias{tf_graph_import_graph_def}
\title{Import a graph definition}
\usage{
tf_graph_import_graph_def(graph_def)
}
\arguments{
\item{graph_def}{A \link[=as_tf_buffer]{tf_buffer} or \code{\link[=raw]{raw()}} vector
containing the}
}
\value{
An object of class 'tf_graph'.
}
\description{
Import a graph definition
}
