% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{tf_register}
\alias{tf_register}
\title{Register a tf vector}
\usage{
tf_register(x, ..., arg = NULL, verbose = FALSE)
}
\arguments{
\item{x}{a tf vector to register.}

\item{...}{additional arguments passed to \code{\link[fdasrvf:time_warping]{fdasrvf::time_warping()}}.}

\item{arg}{the arg vector.}

\item{verbose}{print progress messages?}
}
\value{
\code{list()} containing: registered tf vector \code{x} and warping functions \code{warp}.
}
\description{
Register a tf vector
}
