% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{tf_register}
\alias{tf_register}
\title{Register a tf vector against a template function}
\usage{
tf_register(x, ..., template = NULL, method = "srvf")
}
\arguments{
\item{x}{a tf vector of functions to register.}

\item{...}{additional arguments passed to further methods.}

\item{template}{a tf vector of a template function to register against.}

\item{method}{the implementation method to choose.}
}
\value{
tf vector of the aligning functions, i.e. the warping functions.
}
\description{
Register a tf vector against a template function
}
\references{
Ramsay, O. J, Hooker, Giles, Graves, Spencer (2009).
\emph{Functional Data Analysis with R and MATLAB},  series Use R!, 1 edition.
Springer New York, New York.
ISBN 978-0-387-98184-0, \doi{10.1007/978-0-387-98185-7}.

Srivastava A, Wu W, Kurtek S, Klassen E, Marron JS (2011).
\dQuote{Registration of Functional Data Using Fisher-Rao Metric.}
1103.3817, \url{https://arxiv.org/abs/1103.3817}.

Tucker JD, Wu W, Srivastava A (2013).
\dQuote{Generative models for functional data using phase and amplitude separation.}
\emph{Computational Statistics & Data Analysis}, \bold{61}, 50-66.
ISSN 0167-9473, \doi{https://doi.org/10.1016/j.csda.2012.12.001}, \url{https://www.sciencedirect.com/science/article/pii/S0167947312004227}.
}
