% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{tf_unwarp}
\alias{tf_unwarp}
\title{Unwarp a tf vector}
\usage{
tf_unwarp(x, warp, ..., keep_arg = FALSE)
}
\arguments{
\item{x}{tf vector of unregistered functions.}

\item{warp}{tf vector of warping functions.}

\item{...}{additional arguments passed to \code{\link[=tfd]{tfd()}}.}

\item{keep_arg}{re-eval on original arg after warping or return (irregular)
tf on warped arg (default)?}
}
\value{
the unwarped tf vector, i.e. the registered functions.
}
\description{
Apply (inverse of given warping functions)/(given aligning functions) to
functional data: \eqn{x(t) \to x(h^{-1}(t)) = x(s)}.
}
\examples{
\dontrun{
# TODO: add examples
tf_unwarp(x, warp)
}
}
