% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{tf_warp}
\alias{tf_warp}
\alias{tf_unwarp}
\title{Warp and unwarp tf vectors}
\usage{
tf_warp(x, warp, ..., keep_arg = FALSE)

tf_unwarp(x, warp, ..., keep_arg = FALSE)
}
\arguments{
\item{x}{tf vector of functions. For \code{tf_warp()}, these should be
registered/aligned functions and unaligned functions for \code{tf_unwarp()}.}

\item{warp}{tf vector of warping functions used for transformation.}

\item{...}{additional arguments passed to \code{\link[=tfd]{tfd()}}.}

\item{keep_arg}{re-evaluate on warped arg values after (un)warping or return
tf on (un)warped arg values (default)?}
}
\value{
\itemize{
\item \code{tf_warp()}: the warped tf vector (unregistered functions)
\item \code{tf_unwarp()}: the unwarped tf vector (registered/aligned functions)
}
}
\description{
\code{tf_warp()} applies warping functions to aligned (registered) functional data
to recover the original unregistered curves: \eqn{x(s) \to x(h(s)) = x(t)}.
Conversely, \code{tf_unwarp()} applies the inverse warping to unregistered data
to obtain aligned (registered) functions: \eqn{x(t) \to x(h^{-1}(t)) = x(s)}.
}
\examples{
# generate some random functions
x <- tf_rgp(10)
# apply random warping
warp <- tf_rgam(10)
reg <- tf_warp(x, warp)
# or register the functions
warp <- tf_register(x)
reg <- tf_warp(x, warp)

# get the warping function from registration
warp <- tf_register(pinch)
plot(warp)

# apply the warping to register the data
pinch_reg <- tf_unwarp(pinch, warp)

# round-trip should be approximately equal to original
isTRUE(all.equal(pinch, tf_warp(pinch_reg, warp), tolerance = 0.02))
}
