% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{tf_warp}
\alias{tf_warp}
\alias{tf_unwarp}
\title{Warp and unwarp tf vectors}
\usage{
tf_warp(x, warp, ..., keep_arg = FALSE)

tf_unwarp(x, warp, ..., keep_arg = FALSE)
}
\arguments{
\item{x}{tf vector of functions. For \code{tf_warp()}, these should be
registered/aligned functions and unaligned functions for \code{tf_unwarp()}.}

\item{warp}{tf vector of warping functions used for transformation.}

\item{...}{additional arguments passed to \code{\link[=tfd]{tfd()}}.}

\item{keep_arg}{re-evaluate on warped arg values after (un)warping or return
tf on (un)warped arg values (default)?}
}
\value{
\itemize{
\item \code{tf_warp()}: the warped tf vector (unregistered functions)
\item \code{tf_unwarp()}: the unwarped tf vector (registered/aligned functions)
}
}
\description{
\code{tf_warp()} applies warping functions to aligned (registered) functional data
to recover the original unregistered curves: \eqn{x(s) \to x(h(s)) = x(t)}.
Conversely, \code{tf_unwarp()} applies the inverse warping to unregistered data
to obtain aligned (registered) functions: \eqn{x(t) \to x(h^{-1}(t)) = x(s)}.
}
\examples{
# generate template
template <- tf_rgp(1, arg = 201L)
# generate random warping functions (monotone, [0, 1] -> [0, 1])
warp <- {
  tmp <- exp(tf_rgp(5))
  tf_integrate(tmp, definite = FALSE) / tf_integrate(tmp)
}
x <- tf_warp(rep(template, 5), warp)
layout(t(1:3))
plot(template); plot(warp, col = 1:5); plot(x, col = 1:5)
# register the functions:
warp_estimate <- tf_register(x)
template_estimate <- tf_unwarp(x, warp_estimate)
layout(t(1:2))
plot(warps, col = 1:5); lines(warp_estimate, lty = 2, col = 1:5)
plot(template_estimate, col = 1:5); lines(template, lty = 2)
}
