% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb-mv.R
\name{tfb_mv}
\alias{tfb_mv}
\alias{tfb_mv.tf_mv}
\alias{tfb_mv.list}
\alias{tfb_mv.default}
\alias{as.tfb_mv}
\alias{as.tfb_mv.default}
\title{Vector-valued functional data in basis representation (\code{f: R -> R^d})}
\usage{
tfb_mv(data, ...)

\method{tfb_mv}{tf_mv}(data, basis = c("spline", "fpc"), ...)

\method{tfb_mv}{list}(data, basis = c("spline", "fpc"), ...)

\method{tfb_mv}{default}(data, basis = c("spline", "fpc"), ...)

as.tfb_mv(data, ...)

\method{as.tfb_mv}{default}(data, ...)
}
\arguments{
\item{data}{a \code{\link[=tfd_mv]{tfd_mv()}} / \code{tfb_mv} object, a (named) \code{list} of univariate
\code{tf} vectors, or anything \code{\link[=tfd_mv]{tfd_mv()}} accepts (it is converted to \code{tfd_mv}
first and then each component is expanded into a basis).}

\item{...}{forwarded to the univariate \code{\link[=tfb]{tfb()}} constructor.}

\item{basis}{spline (default) or fpc basis, see \code{\link[=tfb]{tfb()}}.}
}
\value{
a \code{tfb_mv} object.
}
\description{
\code{tfb_mv} is the basis-representation analogue of \code{\link[=tfd_mv]{tfd_mv()}}: it bundles \code{d}
univariate \code{\link[=tfb]{tfb()}} vectors (one per output dimension / component) into a
single vctrs vector of vector-valued functions \eqn{f: \mathbb{R} \to
\mathbb{R}^d}. Each component is fitted independently with the usual
univariate \code{\link[=tfb]{tfb()}} machinery (spline or FPC basis), so all of its arguments
(\code{k}, \code{bs}, \code{penalized}, \code{basis}, ...) apply per component.
}
\examples{
traj <- tfd_mv(list(x = tf_rgp(5), y = tf_rgp(5)))
tb <- tfb_mv(traj, k = 7, verbose = FALSE)
tb
tf_ncomp(tb)
}
\seealso{
Other tf_mv-class: 
\code{\link{plot.tf_mv}()},
\code{\link{tf_ncomp}()},
\code{\link{tfd_mv}}
}
\concept{tf_mv-class}
