% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_id}
\alias{unique_id}
\title{Make syntactically valid unique names}
\usage{
unique_id(x)
}
\arguments{
\item{x}{any input}
}
\value{
\code{x} turned into a list.
}
\description{
See above.
}
\seealso{
Other tidyfun developer tools: 
\code{\link{ensure_list}()},
\code{\link{prep_plotting_arg}()}
}
\concept{tidyfun developer tools}
